backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-jose[cryptography]
requests
boto3
tree-sitter
tree-sitter-python
tree-sitter-java
tree-sitter-javascript
tree-sitter-html
tree-sitter-typescript
tree-sitter-css
tree-sitter-c
tree-sitter-cpp
tree-sitter-go
tree-sitter-kotlin
langchain
langchain-community
langchain-openai
openai
mistralai
